+++ arch/arm/dev/ih_at91piob.c
@@ -91,8 +91,8 @@ void PortIoIrqEntry(void)
 static int PortIoIrqCtl(int cmd, void *param)
 {
     int rc = 0;
-    u_int *ival = (u_int *)param;
-    int enabled = inr(AIC_IMR) & _BV(PIOB_ID);
+    unsigned int *ival = (unsigned int *)param;
+    int_fast8_t enabled = inr(AIC_IMR) & _BV(PIOB_ID);
 
     /* Disable interrupt. */
     if (enabled) {
@@ -124,7 +124,7 @@ static int PortIoIrqCtl(int cmd, void *param)
         break;
     case NUT_IRQCTL_GETMODE:
         {
-            u_int val = inr(AIC_SMR(PIOB_ID)) & AIC_SRCTYPE;
+            unsigned int val = inr(AIC_SMR(PIOB_ID)) & AIC_SRCTYPE;
             if (val == AIC_SRCTYPE_INT_LEVEL_SENSITIVE || val == AIC_SRCTYPE_EXT_HIGH_LEVEL) {
                 *ival = NUT_IRQMODE_LEVEL;
             } else  {
@@ -149,7 +149,7 @@ static int PortIoIrqCtl(int cmd, void *param)
         break;
 #ifdef NUT_PERFMON
     case NUT_IRQCTL_GETCOUNT:
-        *ival = (u_int)sig_PIOB.ir_count;
+        *ival = (unsigned int)sig_PIOB.ir_count;
         sig_PIOB.ir_count = 0;
         break;
 #endif
